+++ ruby/test/test_domain_event_encryption.rb
@@ -3,9 +3,8 @@
 require "test_helper"
 require "securerandom"
 require "json"
-require_relative "../lib/lime_pie"
 
-class TestLimePie < Minitest::Test
+class TestSymmetricEncryption < Minitest::Test
   def test_encrypts_domain_event
     domain_event = File.read(
       "../domain-events/bounded-contexts/studentenverwaltung/events/telefonnummer-geaendert/event-specs/default.json"
@@ -13,7 +12,7 @@ class TestLimePie < Minitest::Test
     raw_number = "+49151234567890"
     named_key = "randomlyNamedKey:#{SecureRandom.hex(16)}"
 
-    encrypted_domain_event = ::LimePie.encrypt_domain_event(JSON(domain_event), named_key)
+    encrypted_domain_event = ::SymmetricEncryption::DomainEvent.encrypt_domain_event(JSON(domain_event), named_key)
     assert !encrypted_domain_event.include?(raw_number)
   end
 
@@ -23,9 +22,9 @@ class TestLimePie < Minitest::Test
     domain_event = File.read(
       "../domain-events/bounded-contexts/kundenverwaltung/events/anrede-geaendert/event-specs/default.json"
     )
-    encrypt_domain_event = ::LimePie.encrypt_domain_event(JSON(domain_event), named_key)
+    encrypt_domain_event = ::SymmetricEncryption::DomainEvent.encrypt_domain_event(JSON(domain_event), named_key)
 
-    decrypted_domain_event = ::LimePie.decrypt_domain_event(encrypt_domain_event, [named_key])
+    decrypted_domain_event = ::SymmetricEncryption::DomainEvent.decrypt_domain_event(encrypt_domain_event, [named_key])
 
     assert decrypted_domain_event["aendert_anrede"]["anrede"] == raw_anrede
   end
